Added a gamesetup function
So the changes I made to main may or may not work. I wrote the code that
way for it to make sense to me. This is sort of based off my original
code that I made before with the "set_up" function. The first part is
made for odd numbered dimensions while the second one is made for even
numbered ones. Forgive me if it is riddled with errors and functions
don't match up. I think we can soon work on the actual battles now.